Default lineHeight and base values
The default "lineHeight" and "base" values provided on the "common" line of the .fnt and .xml output now dynamically correspond to the height of the characters. Prior to this, the values were hard-coded to a default of "32" and "25", meaning that small fonts with heights of 8 would be rendered far off the page.